RE ORG
webpack was importing all modules when just one was imported in a content script, so each content script action was occuring multiple times. Couldn't figure out how to change the webpack/Babel config to fix this so instead put all highlighting related functions into the highlighter.js file. It's a bit unweildy, but performance has improved. In the future, individual conext scripts CAN be created, and can also use exports, as long as the file that imports them will use all the imports, and that no other file is importing it.